solving post merge conflicts

+++ src/main/java/fr/unice/polytech/equipe/j/restaurant/OrderPrice.java
@@ -1,6 +1,6 @@
 package fr.unice.polytech.equipe.j.restaurant;
 import java.util.Map;
 
-public record OrderPrice(Map<MenuItem, Integer> newPrices, double totalPrice, String description){
+public record OrderPrice(Map<MenuItem, Double> newPrices, double totalPrice, String description){
 
 }
